requirements.txt
numpy
pandas
matplotlib
pillow
opencv-python
tqdm
pyyaml
